Migrate Remote Google Fonts to font self-hosting
Remote font loading without prior explicit user consent is not GDPR
compliant, as the client-side request to Google makes usage of the client
IP subject to Google's own terms and conditions.

To keep the remote font loading approach, a user consent system should
be in place. Instead of complicating our approach by implementing such
system and in order to avoid liability issues with GDPR violations,
the new approach is to simply self host the currently required fonts
as their licensing allows it.